examples: collapse the two k8s examples into a single examples/k8s/main.rs
Merge k8s_node and k8s_heartbeat into one multi-file example whose cargo
target is the directory name (cargo run --example k8s). The heartbeat/logging
demo behaviour is kept but fenced with '--- demo ---' markers so it can be
dropped for a bare production node. This removes the EXAMPLE build arg from the
Dockerfile (one binary to build) and co-locates the whole turnkey example
under examples/k8s/. Verified: example targets are now just demo + k8s, fmt and
clippy --all-features pass, and a two-node run still converges.
