+++ Cargo.toml
@@ -58,15 +58,13 @@ dns-hickory = []
 # No deps; not part of the supported public API.
 internal-testing = []
 
+# The turnkey Kubernetes example. It is a multi-file example (`examples/k8s/main.rs`), so its
+# target is named after the directory: `cargo run --example k8s`. The metrics/readiness endpoint
+# (the Kubernetes probes target `/metrics`) is built in via the feature below. The rest of
+# `examples/k8s/` (manifests, kind playground, Dockerfile) is deployment scaffolding.
 [[example]]
-name = "k8s_node"
-# Builds the metrics/readiness endpoint into the binary; the Kubernetes probes target `/metrics`.
-required-features = ["metrics-prometheus"]
-
-[[example]]
-name = "k8s_heartbeat"
-# Same as `k8s_node`, plus a periodic per-pod heartbeat write and a hook that logs each key as it
-# arrives, so reconciliation between pods is visible in `kubectl logs`. Used by `examples/k8s/kind/`.
+name = "k8s"
+path = "examples/k8s/main.rs"
 required-features = ["metrics-prometheus"]
 
 [dependencies]
