+++ examples/k8s/main.rs
@@ -6,23 +6,38 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-//! A Kubernetes-native reconcile node.
+//! A Kubernetes-native reconcile node — the example behind `examples/k8s/`.
 //!
 //! It binds to its pod IP, derives a stable node identity from the pod name, discovers its peers by
 //! resolving a headless `Service` over DNS, and exposes a `/metrics` endpoint for the kubelet
 //! probes. Everything is taken from the environment (the Kubernetes downward API + a ConfigMap /
-//! Secret), so the same image works for every replica. See `examples/k8s/base/` for the manifests.
+//! Secret), so the same image works for every replica. See `examples/k8s/base/` for the manifests
+//! and `examples/k8s/kind/` for a local playground.
 //!
-//! Run it with the metrics endpoint built in:
+//! To make reconciliation *observable*, it also runs a small **demo** behaviour (clearly fenced
+//! below with `--- demo ---` markers — delete those two blocks for a bare production node):
+//!
+//!   * every few seconds the pod writes one key for itself — `heartbeat/<pod-name>` — with the
+//!     current time as its value;
+//!   * a pre-insert hook logs each key the first time it appears in the local store.
+//!
+//! Because the hook fires for updates merged *from peers* too, `kubectl logs reconcile-2` shows the
+//! node learning about `heartbeat/reconcile-0`, `heartbeat/reconcile-1`, … as gossip reconciles the
+//! cluster — i.e. you watch the replicas converge in real time, straight from the logs.
+//!
+//! Run it locally:
 //!
 //! ```sh
-//! cargo run --release --example k8s_node --features metrics-prometheus
+//! POD_IP=127.0.0.1 RECONCILE_DNS_NAME=localhost \
+//!   cargo run --release --example k8s --features metrics-prometheus
 //! ```
 
 use std::collections::hash_map::DefaultHasher;
+use std::collections::HashSet;
 use std::hash::{Hash, Hasher};
 use std::net::{IpAddr, SocketAddr};
-use std::time::Duration;
+use std::sync::{Arc, Mutex};
+use std::time::{Duration, SystemTime, UNIX_EPOCH};
 
 use tracing::{info, warn};
 
@@ -88,9 +103,15 @@ async fn main() {
     let metrics_addr: SocketAddr = optional("RECONCILE_METRICS_ADDR", "0.0.0.0:9000")
         .parse()
         .expect("RECONCILE_METRICS_ADDR must be host:port");
+    // How often this pod refreshes its own heartbeat key.
+    let heartbeat_interval = Duration::from_secs(
+        optional("RECONCILE_HEARTBEAT_SECS", "5")
+            .parse()
+            .expect("RECONCILE_HEARTBEAT_SECS must be an integer"),
+    );
 
-    // Build the config. Note: NO `with_net` is declared — in Kubernetes, peers are found purely
-    // through DNS-resolved pod IPs, so the engine's random per-network probing is left disabled.
+    // Build the config. No `with_net` is declared: in Kubernetes, peers come purely from DNS, so
+    // the engine's random per-network probing is left disabled.
     let mut config = Config::default()
         .with_port(port)
         .with_listen_addr(pod_ip)
@@ -115,6 +136,38 @@ async fn main() {
         .with_discovery_interval(discovery_interval)
         .with_discovery_miss_threshold(miss_threshold);
 
-    info!(%pod_ip, %pod_name, port, "reconcile k8s node starting; discovering peers over DNS");
+    // --- demo (remove for a bare production node) -------------------------------------------
+    // Log each key the first time it lands in the local store. The pre-insert hook fires for local
+    // writes AND for updates merged from peers, so this is exactly where reconciliation becomes
+    // visible: the lines for OTHER pods' heartbeats only appear once gossip has delivered them.
+    let seen: Arc<Mutex<HashSet<String>>> = Arc::new(Mutex::new(HashSet::new()));
+    {
+        let seen = seen.clone();
+        store.add_pre_insert(move |key: &String, value| {
+            // `value.1` is `None` for a tombstone (a delete); only announce live keys.
+            if value.1.is_some() && seen.lock().unwrap().insert(key.clone()) {
+                info!(%key, "store now holds this key (local write or reconciled from a peer)");
+            }
+        });
+    }
+
+    // Refresh this pod's own heartbeat key on a fixed cadence, on a clone of the (Arc-backed) store
+    // handle. `run` below owns its own handle for the gossip loop.
+    let heartbeat_store = store.clone();
+    let heartbeat_key = format!("heartbeat/{pod_name}");
+    tokio::spawn(async move {
+        let mut ticker = tokio::time::interval(heartbeat_interval);
+        loop {
+            ticker.tick().await;
+            let now = SystemTime::now()
+                .duration_since(UNIX_EPOCH)
+                .unwrap_or_default()
+                .as_secs();
+            heartbeat_store.insert(heartbeat_key.clone(), now.to_string());
+        }
+    });
+    // --- end demo ---------------------------------------------------------------------------
+
+    info!(%pod_ip, %pod_name, port, "reconcile k8s node starting; writing heartbeats; discovering peers over DNS");
     store.run().await;
 }
